Update ESPStringTemplate.h
Removing usage of string helper, and using PROGMEM related macros directly. This has removed type checking, and caused the addition of extra functions that are capable of handling data that is stored in progmem.

Also added is the ability to use token string pairs with const char*s directly. This is simpler, faster, and makes sense when not using an array of TokenStringPairs.